% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stationary.R
\docType{data}
\name{data_stationary}
\alias{data_stationary}
\title{Sample data for illustration of stationary coefficient}
\format{
A data frame with 1000 rows and 4 variables, there relation ship follows the following equation
\deqn{Y_t = 40+0.5Y_{t-1}-1.5X_t-0.5X_{t-1}-C_t+v_t, v_t \sim \text{i.i.d }\mathcal{N}(0,1)}
\describe{
  \item{Date}{A simulated date}
  \item{y}{outcome, with data generating process as above}
  \item{c}{continous covarite}
  \item{x}{continous treatment}
}
}
\usage{
data_stationary
}
\description{
Sample data for illustration of stationary coefficient
}
\keyword{datasets}
