% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_SSM.R
\name{explore_SSM}
\alias{explore_SSM}
\title{Exploration function as initial guess}
\usage{
explore_SSM(res = res, plotFlag = T)
}
\arguments{
\item{res}{result subject, generated from \code{shrinkTVP} function}

\item{plotFlag}{whether we need to plot the result, default value is T}
}
\value{
Tables including non-stationary coefficient, possible change points, fitted ARIMA model
}
\description{
Exploration function as initial guess
}
