% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.missing_index.R
\name{generate.missing_index}
\alias{generate.missing_index}
\title{A function for generating}
\usage{
generate.missing_index(type, param, n, printFlag = T)
}
\arguments{
\item{type}{three missing type MCAR,MAR,MNAR}

\item{param}{must be a list and \code{param$data} must be a data.frame}

\item{n}{the length of original data}

\item{printFlag}{print intermediate results and plots, default is T}
}
\value{
A list of result, containing the missing index
}
\description{
A function for generating
}
\examples{
\dontrun{generate.missing_index(type = "MNAR", n=length(data$y),param = list(data = data.frame(y=data$y,x=data$x), MNAR.type = "increasing", coeff = c(0.5,-1,0), MNAR.drawplot = c(TRUE, "y")))$missing_index}


}
