% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMimpute_unanimous_cpts.R
\name{SSMimpute_unanimous_cpts}
\alias{SSMimpute_unanimous_cpts}
\title{Main function, state space model multiple imputation with unanimous changepoints}
\usage{
SSMimpute_unanimous_cpts(
  data_ss_ori,
  formula_var,
  ss_param_temp,
  initial_imputation_option = "StructTS",
  estimate_convergence_cri = 0.01,
  lik_convergence_cri = 0.01,
  stepsize_for_newpart = 1/3,
  max_iteration = 100,
  cpt_learning_param = list(cpt_method = "mean", burnin = 1/10, mergeband = 20,
    convergence_cri = 15),
  cpt_initial_guess_option = "ignore",
  dlm_option = "smooth",
  m = 5,
  seed = 1,
  printFlag = T
)
}
\arguments{
\item{data_ss_ori}{contains all information, and only selected variables in formula_var enters the state space model}

\item{formula_var}{select variables from \code{data_ss} into the statespace model}

\item{ss_param_temp}{a list of parameters to set up state-space model
\itemize{
\item{\code{m0}: }{initial values for states}
\item{\code{C0}: }{initial values for variance of states}
\item{\code{inits}: }{initial values for the estimating of all NA terms, via maximizing likelihood}
\item{\code{AR1_coeffi}: }{variables, whose coefficient is a AR(1) process; if none, then is NULL}
\item{\code{rw_coeffi}: }{variables, whose coefficient is a random walk process;if none, then is NULL}
\item{\code{w_cp_param}: }{variables, whose coefficients are periodic fixed (may shift to other levels over time, but fixed within periods), it contains a list of parameters for each variable whose coefficient level shifts to different values. Details for variable: \code{variable} : the name of the variable (Required); \code{segments} : how many segments of constant coefficient (Required); \code{changepoints} :the corresponding changepoints for the separated segments, it can either be specified by the user or automatically inferred, \code{fixed_cpts} only exist when \code{changepoints} exists }
\item{\code{v_cp_param}: }{information about periodic observational variance V (may decrease or increase over time, but fixed within periods), it contains a list of parameters for each variable whose coefficient level shifts to different values. Details for variable: \code{segments} : how many segments of constant coefficient (Required); \code{changepoints} : the corresponding changepoints for the separated segments, it can either be specified by the user or automatically inferred, \code{fixed_cpts} only exist when \code{changepoints} exists}
}}

\item{initial_imputation_option}{for the first iteration of imputing missing y, choose StructTS or others, and can't be "ignore"}

\item{estimate_convergence_cri}{critic value for convergence check, default 0.01}

\item{lik_convergence_cri}{critic value for convergence check, default 0.01}

\item{stepsize_for_newpart}{stepsize specified, default 1/3}

\item{max_iteration}{max iteration, default 100}

\item{cpt_learning_param}{a list of variable for change point learning
\itemize{
\item{\code{cpt_method}: }{either "mean" or "meanvar"}
\item{\code{burnin}: }{a positive number in (0,1)}
\item{\code{mergeband}: }{a positive integer}
\item{\code{AR1_coeffi}: }{variables, whose coefficient is a AR(1) process; if none, then is NULL}
\item{\code{rw_coeffi}: }{variables, whose coefficient is a random walk process;if none, then is NULL}
\item{\code{w_cp_param}: }{variables, whose coefficients are periodic fixed (may shift to other levels over time, but fixed within periods)}
\item{\code{v_cp_param}: }{information about periodic observational variance V (may decrease or increase over time, but fixed within periods)}
}}

\item{cpt_initial_guess_option}{option for initially learning cpts in preparation period}

\item{dlm_option}{choose between smooth or filter}

\item{m}{number of draws for multiple imputation}

\item{seed}{random seed}

\item{printFlag}{whether we need to print the Flag plots.}
}
\value{
A list of result
}
\description{
Main function, state space model multiple imputation with unanimous changepoints
}
